yolov8-images/code/requirements.txt
opencv-python
torchvision
seaborn
ultralytics
omegaconf==2.3.0
PIL